pyproject.toml
[project]
name = "rishta-bot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "uvicorn>=0.35.0",
]
